Does this simplify review by only breaking out within in this pr?
