+++ packages/driver/src/cy/commands/querying/index.ts
@@ -1,15 +1,7 @@
-import * as Contains from './contains'
-import * as Focused from './focused'
-import * as Get from './get'
-import * as Root from './root'
-import * as Shadow from './shadow'
+import * as Querying from './querying'
 import * as Within from './within'
 
 export {
-  Contains,
-  Focused,
-  Get,
-  Root,
-  Shadow,
+  Querying,
   Within,
 }
